safari-600.7-branch/LayoutTests/editing/selection/move-by-line-006.html

<!DOCTYPE html>
<html>
<body>
<p>To manually run the test, put the cursor at the beginning of the next paragraph,
and then press the down arrow key. The caret should go to the next line,
not the beginning of the dashed section.</p>

<div id="result" style="font-family: Ahem; font-size: 16px; color: #fcc; width:400px;">
aaaa aaaa
<span style="border-style: dashed; background-color: #000">
aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa
aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa</span>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
document.designMode = 'on';

var selection = window.getSelection();
var result = document.getElementById('result');
selection.setPosition(result, 0);
selection.modify('move', 'forward', 'line');

Markup.dump('result');
</script>
</body>
</html>
