LayoutTests/editing/selection/move-by-line-006-expected.txt

| "\naaaa aaaa\n"
| <span>
|   style="border-style: dashed; background-color: #000"
|   "\naaaa aaaa aaaa <#selection-caret>aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa\naaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa aaaa"
| "\n"
